Reimplemented defaults => placeholders
Previously defaults could only be static images.

I wanted to support having autogenerated avatars that are unique for every user. In order to support this I "needed" to be able to change the defaults behavior.

For this reason I renamed defaults to placeholders and moved it out into it's own file that anyone can replace.

So now in my app I will write my own Placeholder and change the placeholder class in the cofig to mine. My placeholder class will then autogenerate an avatar for the model and attach it to the model.